# ios/Sessioni Sotterranee/Sessioni Sotterranee/openapi.yaml
openapi: 3.1.0
info:
  title: ""
  version: ""
paths:
  /api/events/:
    get:
      description: Get paginated upcoming events
      parameters:
        - in: query
          name: limit
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 0
          style: form
        - in: query
          name: offset
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 0
          style: form
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        "500":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
  /api/events/{id}:
    get:
      description: Get an event by id
      parameters:
        - in: path
          name: id
          description: Event id
          required: true
          schema:
            type: integer
            format: int32
          style: simple
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        "500":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
  /api/events/forme:
    get:
      description: Get paginated attending and interested events for the current user
      parameters:
        - in: query
          name: limit
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 0
          style: form
        - in: query
          name: offset
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 0
          style: form
        - in: header
          name: Authorization
          description: Jwt Bearer token
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsForUser'
        "500":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
  /api/users/set_username:
    patch:
      description: complete user registration
      parameters:
        - in: header
          name: Authorization
          description: Jwt Bearer token
          required: true
          schema:
            type: string
          style: simple
        - in: query
          name: username
          required: true
          schema:
            type: string
          style: form
      responses:
        "200":
          description: no content
        "500":
          description: ""
        "401":
          description: ""
        "404":
          description: ""
  /ws:
    get:
      responses:
        "101":
          description: websocket upgrade
          headers:
            connection:
              style: simple
              schema:
                type: string
                enum:
                  - upgrade
                const: upgrade
                example: upgrade
            upgrade:
              style: simple
              schema:
                type: string
                enum:
                  - websocket
                const: websocket
                example: websocket
            sec-websocket-key:
              style: simple
              schema:
                type: string
            sec-websocket-protocol:
              style: simple
              schema:
                type: string
components:
  schemas:
    CompleteRegistrationRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
    Event:
      type: object
      required:
        - date
        - id
        - title
      properties:
        date:
          type: string
          format: date-time
        description:
          type:
            - string
            - "null"
        genre:
          type:
            - string
            - "null"
        id:
          type: integer
          format: int32
        image:
          type:
            - string
            - "null"
        location:
          type:
            - string
            - "null"
        title:
          type: string
        url:
          type:
            - string
            - "null"
    EventsForUser:
      type: object
      required:
        - attending
        - interested
      properties:
        attending:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        interested:
          type: array
          items:
            $ref: '#/components/schemas/Event'
    PaginationQuery:
      type: object
      required:
        - limit
        - offset
      properties:
        limit:
          type: integer
          format: uint64
          minimum: 0
        offset:
          type: integer
          format: uint64
          minimum: 0
    SelectEvent:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int32